fix: 당첨금 자료형을 long으로 변경 및 테스트 추가

+++ src/main/java/Application.java
@@ -32,7 +32,7 @@ public class Application {
         WinningLotto winningLotto = WinningLotto.from(wins, bonusBall);
         List<Rank> ranks = lottoResult.calculateRanks(passiveLotto, winningLotto);
 
-        int winPrice = winningRate.calculateWinPrice(ranks);
+        long winPrice = winningRate.calculateWinPrice(ranks);
         double rate = winningRate.calculateRate(winPrice, purchaseAmount);
 
         ResultView.printResult(ranks, rate);
